+++ R/spgautor.R
@@ -59,9 +59,10 @@
 #'   specifying the partition factor.  The partition factor assumes observations
 #'   from different levels of the partition factor are uncorrelated.
 #' @param W Weight matrix specifying the neighboring structure used.
-#'   Not required if \code{data} is an \code{sf} polygon object,
+#'   Not required if \code{data} is an \code{sf} object wtih \code{POLYGON} geometry,
 #'   as \code{W} is calculated internally using queen contiguity. If calculated internally,
-#'   \code{W} is computed using \code{sf::st_intersects()}.
+#'   \code{W} is computed using \code{sf::st_intersects()}. Also not required if \code{data}
+#'   is an \code{sf} object with \code{POINT} geometry as long as \code{cutoff} is specified.
 #' @param row_st A logical indicating whether row standardization be performed on
 #'   \code{W}. The default is \code{TRUE}.
 #' @param M \code{M} matrix satisfying the car symmetry condition. The car
@@ -75,6 +76,10 @@
 #'   or given as a vector or one-column matrix assumed to be the diagonal.
 #' @param range_positive Whether the range should be constrained to be positive.
 #'   The default is \code{TRUE}.
+#' @param cutoff The numeric, distance-based cutoff used to determine \code{W}
+#'   when \code{W} is not specified. For an \code{sf} object with \code{POINT} geometry,
+#'   two locations are considered neighbors if the distance between them is less
+#'   than or equal to \code{cutoff}.
 #' @param ... Other arguments to [stats::optim()].
 #'
 #' @details The spatial generalized linear model for areal data
@@ -223,7 +228,7 @@
 #' @references
 #' McCullagh P. and Nelder, J. A. (1989) \emph{Generalized Linear Models}. London: Chapman and Hall.
 spgautor <- function(formula, family, data, spcov_type, spcov_initial, dispersion_initial,
-                     estmethod = "reml", random, randcov_initial, partition_factor, W, row_st = TRUE, M, range_positive = TRUE, ...) {
+                     estmethod = "reml", random, randcov_initial, partition_factor, W, row_st = TRUE, M, range_positive = TRUE, cutoff, ...) {
 
   # set car as default if nothing specified
   if (missing(spcov_type) && missing(spcov_initial)) {
@@ -315,11 +320,15 @@ spgautor <- function(formula, family, data, spcov_type, spcov_initial, dispersio
     partition_factor <- NULL
   }
 
+  if (missing(cutoff)) {
+    cutoff <- NULL
+  }
+
   # get data object
   data_object <- get_data_object_spgautor(
     formula, family, data, spcov_initial,
     estmethod, W, M, random, randcov_initial,
-    partition_factor, row_st, range_positive, ...
+    partition_factor, row_st, range_positive, cutoff, ...
   )
 
   cov_est_object <- cov_estimate_laploglik_spgautor(data_object, formula,